accordion ui updates, adding container views
