+++ app/components/ContainerCard.jsx
@@ -12,7 +12,7 @@ const ContainerCard = ({ container }) => {
       className="border-none"
       shadow="sm"
       classNames={{
-        root: "cursor-pointer hover:brightness-90 aspect-[2.5/1] !p-2",
+        root: "cursor-pointer hover:brightness-90 aspect-[2.5/1] !p-4",
       }}
       styles={{
         root: {
