+++ app/items/page.jsx
@@ -6,15 +6,11 @@ import { fetcher } from "../lib/fetcher";
 import { useContext } from "react";
 import { sortObjectArray } from "../lib/helpers";
 import ItemGrid from "../components/ItemGrid";
-import { useViewportSize } from "@mantine/hooks";
-import { ScrollArea } from "@mantine/core";
 import { FilterContext } from "./layout";
 import Loading from "../components/Loading";
 
 const Page = ({ searchParams }) => {
   const query = searchParams?.query || "";
-  const { height } = useViewportSize();
-  const maxHeight = height * 0.72;
   const { data, isLoading, error } = useSWR(
     `/items/api?search=${query}`,
     fetcher
@@ -32,21 +28,14 @@ const Page = ({ searchParams }) => {
   const itemsToShow = selected?.length ? filtered : data?.items;
 
   return (
-    <>
-      <ScrollArea.Autosize
-        mah={maxHeight}
-        classNames={{
-          viewport: "pb-16",
-        }}
-      >
-        <ItemGrid desktop={3}>
-          {sortObjectArray(itemsToShow)?.map((item) => {
-            return <ItemCard key={item.name} item={item} showLocation />;
-          })}
-        </ItemGrid>
-      </ScrollArea.Autosize>
+    <div className="pb-12">
+      <ItemGrid desktop={3}>
+        {sortObjectArray(itemsToShow)?.map((item) => {
+          return <ItemCard key={item.name} item={item} showLocation />;
+        })}
+      </ItemGrid>
       <NewItem data={data} opened={opened} close={close} />
-    </>
+    </div>
   );
 };
